requirements.txt
firebase-admin==6.2.0
python-dotenv==1.0.1
pandas==2.2.1
matplotlib==3.8.3
notebook==7.1.1
protobuf==4.25.3